+++ spyder/plugins/profiler/plugin.py
@@ -1,12 +1,8 @@
 # -*- coding: utf-8 -*-
-# -----------------------------------------------------------------------------
-# Copyright (c) 2009- Spyder Project Contributors
-# Based on p_pylint.py by Pierre Raybaut
 #
-# Distributed under the terms of the MIT License
+# Copyright © Spyder Project Contributors
+# Licensed under the terms of the MIT License
 # (see spyder/__init__.py for details)
-# -----------------------------------------------------------------------------
-
 
 """Profiler Plugin."""
 
@@ -14,51 +10,24 @@
 import os.path as osp
 
 # Third party imports
-from qtpy.QtCore import Qt
-from qtpy.QtWidgets import QGroupBox, QLabel, QVBoxLayout
+from qtpy.QtWidgets import QVBoxLayout
 
 # Local imports
 from spyder.config.base import _
 from spyder.config.gui import is_dark_interface
 from spyder.api.plugins import SpyderPluginWidget
-from spyder.api.preferences import PluginConfigPage
 from spyder.preferences.runconfig import get_run_configuration
 from spyder.utils import icon_manager as ima
 from spyder.utils.qthelpers import create_action
-from .widgets.profilergui import (ProfilerWidget, is_profiler_installed)
+from spyder.plugins.profiler.confpage import ProfilerConfigPage
+from spyder.plugins.profiler.widgets.profilergui import (ProfilerWidget,
+                                                         is_profiler_installed)
 
 
 if is_dark_interface():
     MAIN_TEXT_COLOR = 'white'
 else:
     MAIN_TEXT_COLOR = '#444444'
-
-
-class ProfilerConfigPage(PluginConfigPage):
-    def setup_page(self):
-        results_group = QGroupBox(_("Results"))
-        results_label1 = QLabel(_("Profiler plugin results "
-                                  "(the output of python's profile/cProfile)\n"
-                                  "are stored here:"))
-        results_label1.setWordWrap(True)
-
-        # Warning: do not try to regroup the following QLabel contents with
-        # widgets above -- this string was isolated here in a single QLabel
-        # on purpose: to fix Issue 863
-        results_label2 = QLabel(ProfilerWidget.DATAPATH)
-
-        results_label2.setTextInteractionFlags(Qt.TextSelectableByMouse)
-        results_label2.setWordWrap(True)
-
-        results_layout = QVBoxLayout()
-        results_layout.addWidget(results_label1)
-        results_layout.addWidget(results_label2)
-        results_group.setLayout(results_layout)
-
-        vlayout = QVBoxLayout()
-        vlayout.addWidget(results_group)
-        vlayout.addStretch(1)
-        self.setLayout(vlayout)
 
 
 class Profiler(SpyderPluginWidget):
